Refactor indexer_api -> api
