+++ medusa/metadata/kodi_12plus.py
@@ -9,7 +9,7 @@ from babelfish import Country
 from medusa import helpers
 from medusa.app import TVDB_API_KEY
 from medusa.helper.common import dateFormat, episode_num
-from medusa.indexers.indexer_api import indexerApi
+from medusa.indexers.api import indexerApi
 from medusa.indexers.indexer_exceptions import IndexerEpisodeNotFound, IndexerSeasonNotFound
 from medusa.indexers.tvdbv2.tvdbv2_api import API_BASE_TVDB
 from medusa.logger.adapters.style import BraceAdapter
