+++ medusa/scene_exceptions.py
@@ -13,7 +13,7 @@ from collections import defaultdict
 import adba
 
 from medusa import app, db, helpers
-from medusa.indexers.indexer_api import indexerApi
+from medusa.indexers.api import indexerApi
 from medusa.indexers.indexer_config import INDEXER_TVDBV2
 from medusa.session.core import MedusaSafeSession
