+++ medusa/scene_numbering.py
@@ -27,7 +27,7 @@ import traceback
 
 from medusa import app, db, logger
 from medusa.helper.exceptions import ex
-from medusa.indexers.indexer_api import indexerApi
+from medusa.indexers.api import indexerApi
 from medusa.scene_exceptions import safe_session
 from medusa.show.show import Show
