+++ views/displayShow.mako
@@ -8,7 +8,7 @@
     from medusa.common import Quality, qualityPresets, statusStrings, Overview
     from medusa.helpers import anon_url
     from medusa.helper.common import pretty_file_size
-    from medusa.indexers.indexer_api import indexerApi
+    from medusa.indexers.api import indexerApi
     from medusa.indexers.indexer_config import mappings
 %>
 <%block name="scripts">
