+++ views/home_massAddTable.mako
@@ -1,7 +1,7 @@
 <%!
     from medusa import app
     from medusa.helpers import anon_url
-    from medusa.indexers.indexer_api import indexerApi
+    from medusa.indexers.api import indexerApi
 %>
 <table id="addRootDirTable" class="defaultTable tablesorter">
     <thead>
