+++ views/schedule.mako
@@ -2,7 +2,7 @@
 <%!
     from medusa import app
     from medusa.helpers import anon_url
-    from medusa.indexers.indexer_api import indexerApi
+    from medusa.indexers.api import indexerApi
     from medusa.indexers.indexer_config import mappings
     from medusa import sbdatetime
     from random import choice
